feat(k8s): Asset scan support (#864)
* feat: kubernetes scanning

* refactor: cosmetics

* feat: make scanner namespace configurable

* fix(helm): install latest grype-server

* refactor: container discoverer implementations

* style: use more descriptive var name for imageInfo

* refactor: change ContainerImageInfo size to int64

* refactor: image size calculation

* test(e2e): add Container Runtime Discovery Server

* feat: merge Pod labels into ContainerInfo

* fix: image pull policy for scanner pod

* test(e2e): fix setting Helm release name

* test(e2e): add missing label to asset

* refactor: image walk method for containerd

* refactor: mk Helm config options as defaulter

* fix: typos

* refactor: assign backoff limit directly

---------

+++ api/models/models.gen.go
@@ -611,7 +611,7 @@ type ContainerImageInfo struct {
 	Os           *string   `json:"os,omitempty"`
 	RepoDigests  *[]string `json:"repoDigests,omitempty"`
 	RepoTags     *[]string `json:"repoTags,omitempty"`
-	Size         *int      `json:"size,omitempty"`
+	Size         *int64    `json:"size,omitempty"`
 }
 
 // ContainerInfo defines model for ContainerInfo.
